+++ src/error.rs
@@ -2,7 +2,6 @@
 //
 // Copyright (c) 2019  Douglas Lau
 //
-
 use std::fmt;
 use std::io;
 
@@ -60,5 +59,38 @@ impl std::error::Error for DecodeError {
 impl From<io::Error> for DecodeError {
     fn from(e: io::Error) -> Self {
         DecodeError::Io(e)
+    }
+}
+
+/// Errors encountered while encoding a GIF file.
+#[derive(Debug)]
+pub enum EncodeError {
+    /// A wrapped I/O error.
+    Io(io::Error),
+    /// [Block](block/enum.Block.html)s arranged in invalid sequence.
+    InvalidBlockSequence,
+}
+
+impl fmt::Display for EncodeError {
+    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            EncodeError::Io(err) => err.fmt(fmt),
+            _ => fmt::Debug::fmt(self, fmt),
+        }
+    }
+}
+
+impl std::error::Error for EncodeError {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        match *self {
+            EncodeError::Io(ref err) => Some(err),
+            _ => None,
+        }
+    }
+}
+
+impl From<io::Error> for EncodeError {
+    fn from(e: io::Error) -> Self {
+        EncodeError::Io(e)
     }
 }